parseInt -> parseFloat

+++ adapters/resolv.ts
@@ -27,11 +27,11 @@ export default {
   },
   points: (data: Record<string, number | string>) => {
     return {
-      total_amount: parseInt(String(data?.total_amount)) || 0,
+      total_amount: parseFloat(String(data?.total_amount)) || 0,
       rank: data?.rank ?? 0,
       total_attributions: data?.total_attributions ?? 0,
     };
   },
   total: (data: { total_amount?: number }) =>
-    parseInt(String(data?.total_amount)) ?? 0,
+    parseFloat(String(data?.total_amount)) ?? 0,
 } as AdapterExport;
